+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed a bug where a refund could not be created if the orderId parameter wasn't provided.
+
 ## [5.2.0] - 2021-08-27
 ### Added
 - Added support for [API manual capture endpoints](https://docs.multisafepay.com/api/#manual-capture-orders)